% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{create_table}
\usage{
create_table(.fit_list, ...)
}
\arguments{
\item{.fit_list}{A list of fits produced using functions of the analyseD
package.}

\item{...}{Additional arguments to be passed to the function.}
}
\description{
Generic function to create HTML tables from lists of results
}
