% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.reglinout.R
\name{create_table.reglinout}
\alias{create_table.reglinout}
\title{create_table.reglinout}
\usage{
\method{create_table}{reglinout}(.fit_list, .annotation = NULL, ...)
}
\arguments{
\item{.fit_list}{A list of fits produced using functions of the analyzeD
package that is sorted by the outcomes.}

\item{.annotation}{An annotation file.}

\item{...}{Additional arguments to be passed to the function.}
}
\description{
Generic function to create HTML tables from lists of results
}
