% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.reglinpre.R
\name{create_table.reglinpre}
\alias{create_table.reglinpre}
\title{create_table.reglinpre}
\usage{
\method{create_table}{reglinpre}(.fit_list, .annotation = NULL, ...)
}
\arguments{
\item{.fit_list}{A list of fits produced using functions of the analyzeD
package that is sorted by the predictors.}

\item{.annotation}{An annotation file.}

\item{...}{Additional arguments to be passed to the function.}
}
\description{
Generic function to create HTML tables from lists of results
}
