% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.reglog.R
\name{create_table.reglog}
\alias{create_table.reglog}
\title{create_table.reglog}
\usage{
\method{create_table}{reglog}(.fit_list, .annotation = NULL, .only_summary = FALSE, ...)
}
\arguments{
\item{.fit_list}{A list of fits produced using functions of the analyzeD
package.}

\item{.annotation}{An annotation file.}

\item{.only_summary}{Boolean indicating whether to only include the summary.}

\item{...}{Additional arguments to be passed to the function.}
}
\description{
Generic function to create HTML tables from lists of results
}
