% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lin_outcomes.R
\name{reg_lin_outcomes}
\alias{reg_lin_outcomes}
\title{reg_lin_outcomes}
\usage{
reg_lin_outcomes(
  .data,
  .outcomes,
  .predictor,
  .covariates,
  .annotation,
  .std_prd,
  .std_cov,
  .summary,
  .interaction,
  ...
)
}
\arguments{
\item{.data}{A data.frame or .mids object.}

\item{.outcomes}{A character vector containing the outcomes.}

\item{.covariates}{A character vector containing covariates.}

\item{.annotation}{A matrix or data.frame of format (name, pname, unit,
short_pname, comment) that contains pretty names for the used variables.}

\item{.std_prd}{If TRUE, predictors are standardized.}

\item{.std_cov}{Character vector of covariates that should be standardized.}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to specify interactions using a list of
character vectors containing the interaction variables, e.g.
list(c("variable1", "variable2"), c("variable2", "variable3")).}

\item{...}{Optional input passed directly to the regression function.}

\item{.predictors}{A character vector containing the predictors.}
}
\description{
This function calculates the regressions while shuffeling
through the provided outcomes.
}
