% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lin_outcomes_summary.R
\name{reg_lin_outcomes_summary}
\alias{reg_lin_outcomes_summary}
\title{reg_lin_outcomes_summary}
\usage{
reg_lin_outcomes_summary(.fit_list, .outcomes, .predictor, .annotation)
}
\arguments{
\item{.fit_list}{The list of fits produced beforehand.}

\item{.outcomes}{A vector containing the outcomes.}

\item{.predictor}{A vector containing the predictor.}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}
}
\description{
This function produces summary tables.
}
