% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lin_predictors.R
\name{reg_lin_predictors}
\alias{reg_lin_predictors}
\title{reg_lin_predictors}
\usage{
reg_lin_predictors(
  .data,
  .predictors,
  .covariates,
  .outcome,
  .annotation,
  .std_prd,
  .std_cov,
  .summary,
  .interaction,
  ...
)
}
\arguments{
\item{.data}{A data.frame.}

\item{.predictors}{A vector containing the predictors.}

\item{.covariates}{A vector containing covariates for each regression.}

\item{.outcome}{A vector containing the outcome.}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}

\item{.std_prd}{If TRUE, predictors are standardized using std(predictor).}

\item{.std_cov}{Input vector of covariates that are standardized using
std(covariate).}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to input interactions.}

\item{...}{Optional input passed to the regression function.}
}
\description{
This function calculates the regressions while shuffeling
through the input predictors.
}
