% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lin_predictors_summary.R
\name{reg_lin_predictors_summary}
\alias{reg_lin_predictors_summary}
\title{reg_lin_predictors_summary}
\usage{
reg_lin_predictors_summary(.fit_list, .outcome, .annotation)
}
\arguments{
\item{.fit_list}{The list of fits produced beforehand.}

\item{.outcome}{A vector containing the outcome.}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}
}
\description{
This function produces summary tables.
}
