% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lin_sort_by_outcomes.R
\name{reg_lin_sort_by_outcomes}
\alias{reg_lin_sort_by_outcomes}
\title{reg_lin_sort_by_outcomes}
\usage{
reg_lin_sort_by_outcomes(
  .data,
  .outcomes,
  .predictors,
  .covariates,
  .annotation,
  .cpus,
  .std_prd,
  .summary,
  .interaction,
  ...
)
}
\arguments{
\item{.data}{A data.frame.}

\item{.outcomes}{A vector containing the outcomes.}

\item{.predictors}{A vector containing the predictors.}

\item{.covariates}{A vector containing covariates for each regression.}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}

\item{.cpus}{Input number of desired cpus to use. Useful only in case of big
datasets and multiple analysis.}

\item{.std_prd}{If TRUE, predictors are standardized using std(predictor).}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to input interactions.}

\item{...}{Optional input passed to the regression function.}
}
\description{
This function calls the actual reg_lin_predictors
function that shuffles through the predictors.
}
