% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_log_predictors.R
\name{reg_log_predictors}
\alias{reg_log_predictors}
\title{reg_log_predictors}
\usage{
reg_log_predictors(
  .data,
  .predictors,
  .covariates,
  .outcome,
  .annotation,
  .std_prd,
  .std_cov,
  .summary,
  .interaction,
  .firth,
  .imputed_predictors,
  .imputed_outcomes,
  ...
)
}
\arguments{
\item{.data}{A data.frame or .mids object.}

\item{.predictors}{A character vector containing the predictors.}

\item{.covariates}{A character vector containing covariates.}

\item{.outcome}{A character vector containing the outcome.}

\item{.annotation}{A matrix or data.frame of format (name, pname, unit,
short_pname, comment) that contains pretty names for the used variables.}

\item{.std_prd}{If TRUE, predictors are standardized.}

\item{.std_cov}{Character vector of covariates that should be standardized.}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to specify interactions using a list of
character vectors containing the interaction variables, e.g.
list(c("variable1", "variable2"), c("variable2", "variable3")).}

\item{.firth}{If TRUE, Firth-correction is used via brglm().}

\item{.imputed_predictors}{If TRUE, cases with imputed predictors are used.}

\item{.imputed_outcomes}{If TRUE, cases with imputed outcomes are used.}

\item{...}{Optional input passed directly to the regression function.}
}
\description{
This function calculates the regressions while shuffeling
through the provided predictors.
}
