% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_log_sort_by_predictors.R
\name{reg_log_sort_by_predictors}
\alias{reg_log_sort_by_predictors}
\title{reg_log_sort_by_predictors}
\usage{
reg_log_sort_by_predictors(
  .data,
  .outcomes,
  .predictors,
  .covariates,
  .annotation,
  .cpus,
  .std_prd,
  .std_cov,
  .summary,
  .interaction,
  .firth,
  ...
)
}
\arguments{
\item{.data}{A data.frame or .mids object.}

\item{.outcomes}{A character vector containing the outcomes.}

\item{.predictors}{A character vector containing the predictors.}

\item{.covariates}{A character vector containing covariates.}

\item{.annotation}{A matrix or data.frame of format (name, pname, unit,
short_pname, comment) that contains pretty names for the used variables.}

\item{.cpus}{Input number of desired cpus to use. Useful only in case of big
datasets and multiple outcomes/predictors.}

\item{.std_prd}{If TRUE, predictors are standardized.}

\item{.std_cov}{Character vector of covariates that should be standardized.}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to specify interactions using a list of
character vectors containing the interaction variables, e.g.
list(c("variable1", "variable2"), c("variable2", "variable3")).}

\item{.firth}{If TRUE, a Firth-corrected version of glm in brglm() is called.}

\item{...}{Optional input passed directly to the regression function.}
}
\description{
This function calls the actual reg_log_outcomes
function that shuffles through the provided outcomes.
}
