% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_lin.R
\name{regression_lin}
\alias{regression_lin}
\title{regression_lin}
\usage{
regression_lin(
  .data,
  .outcomes = NULL,
  .predictors = NULL,
  .covariates = NULL,
  .annotation = NULL,
  .subset = NULL,
  .cpus = 1,
  .sort_by = "outcomes",
  .std_prd = FALSE,
  .summary = FALSE,
  .interaction = NULL,
  ...
)
}
\arguments{
\item{.data}{A data.frame or .mids object.}

\item{.outcomes}{A vector containing the outcomes.}

\item{.predictors}{A vector containing the predictors.}

\item{.covariates}{A vector containing covariates for each regression.}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}

\item{.subset}{Can be used to internally subset the data. Use .subset =
"variable == 'x'" to subset data.}

\item{.cpus}{Input number of desired cpus to use. Useful only in case of big
datasets and multiple analysis.}

\item{.sort_by}{A character string that indicates either to sort the analyses
by "outcomes" or by "predictors".}

\item{.std_prd}{If TRUE, predictors are standardized using std(predictor).}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to input interactions.}

\item{...}{Optional input passed to the regression function.}
}
\value{
A list containing individual results tables for each regression
analysis.
}
\description{
Calculates linear regression models using lm(). Designed to make
analysis with multiple predictors / outcomes easy.
}
\examples{
-
}
