% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_lin_by_predictors.R
\name{regression_lin_by_predictors}
\alias{regression_lin_by_predictors}
\title{regression_lin_by_predictors}
\usage{
regression_lin_by_predictors(
  .data,
  .outcomes,
  .predictors,
  .covariates,
  .annotation,
  .cpus,
  .std_prd,
  .summary,
  .interaction,
  ...
)
}
\arguments{
\item{.data}{A data.frame}

\item{.outcomes}{A vector containing the outcomes}

\item{.predictors}{A vector containing the predictors}

\item{.covariates}{A vector containing covariates for each regression}

\item{.annotation}{A matrix or data.frame in the annotation format (name,
pname, unit, short_pname, comment) that contains pretty names for the used
variables and their dummy variables.}

\item{.cpus}{Input number of desired cpus to use. Useful only in case of big
datasets and multiple analysis.}

\item{.std_prd}{If TRUE, predictors are standardized using std(predictor).}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to input interactions.}

\item{.subset}{Can be used to internally subset the data. Use .subset =
"variable == 'x'" to subset data.}

\item{.sort_by}{A character string that indicates either to sort the analyses
by "outcomes" or by "predictors".}
}
\description{
This function calls the actual regression_lin_outcomes
function that shuffles through the outcomes
}
