% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_log.R
\name{regression_log}
\alias{regression_log}
\title{regression_log}
\usage{
regression_log(
  .data,
  .outcomes = NULL,
  .predictors = NULL,
  .covariates = NULL,
  .annotation = NULL,
  .subset = NULL,
  .cpus = 1,
  .sort_by = "outcomes",
  .std_prd = FALSE,
  .std_cov = NULL,
  .summary = FALSE,
  .interaction = NULL,
  .firth = FALSE,
  .imputed_predictors = FALSE,
  .imputed_outcomes = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data.frame or .mids object.}

\item{.outcomes}{A character vector containing the outcomes.}

\item{.predictors}{A character vector containing the predictors.}

\item{.covariates}{A character vector containing covariates.}

\item{.annotation}{A matrix or data.frame of format (name, pname, unit,
short_pname, comment) that contains pretty names for the used variables.}

\item{.subset}{Can be used to internally subset the data. Use .subset =
"variable == 'x'" to subset data.}

\item{.cpus}{Input number of desired cpus to use. Useful only in case of big
datasets and multiple outcomes/predictors.}

\item{.sort_by}{A character string that indicates either to sort the analyses
by "outcomes" or by "predictors".}

\item{.std_prd}{If TRUE, predictors are standardized.}

\item{.std_cov}{Character vector of covariates that should be standardized.}

\item{.summary}{If TRUE, an additional summary of all analyses is returned.}

\item{.interaction}{Can be used to specify interactions using a list of
character vectors containing the interaction variables, e.g.
list(c("variable1", "variable2"), c("variable2", "variable3")).}

\item{.firth}{If TRUE, a Firth-corrected version of glm in brglm() is called.}

\item{.imputed_predictors}{If TRUE, cases with imputed predictors are used.}

\item{.imputed_outcomes}{If TRUE, cases with imputed outcomes are used.}

\item{...}{Optional input passed directly to the regression function.}
}
\value{
A list containing individual results tables for each regression
analysis.
}
\description{
Calculates logistic regression models using glm()
or brglm() / brglm2.
Designed to make analysis with multiple predictors / outcomes easy and quick.
}
\examples{
-
}
